优化cellID生成逻辑，减少哈希碰撞，附带碰撞的test

+++ internal/model/rider.go
@@ -5,7 +5,7 @@ type Rider struct {
 	X      int
 	Y      int
 	OnLine bool
-	CellID int
+	CellID int64
 	Count  int64
 }
